Combat Selection Indicator and Damage Popup
Implementing a small selection tool for showing the target selected and a damage popup. Issue with damage popup is that it triggers after the ability animation and not with so that will need to be addressed. Also had to refactor the ability animation flow to allow for things like damage popup to go through.